requirements.txt
requests==2.31.0
pillow==10.1.0
beautifulsoup4==4.12.2
huggingface_hub==0.19.4